moved EnvCapture to its own file
