+++ PolyEngine/RenderingDevice/OpenGL/Src/Proxy/GLCubemapDeviceProxy.cpp
@@ -1,6 +1,6 @@
 #include "Proxy/GLCubemapDeviceProxy.hpp"
 #include "Proxy/GLTextureDeviceProxy.hpp"
-#include "Proxy/GLUtils.hpp"
+#include "Common/GLUtils.hpp"
 
 using namespace Poly;
