+++ PolyEngine/RenderingDevice/OpenGL/Src/Proxy/GLMeshDeviceProxy.cpp
@@ -1,6 +1,6 @@
 #include "Proxy/GLMeshDeviceProxy.hpp"
 #include "Math/Vector3f.hpp"
-#include "Proxy/GLUtils.hpp"
+#include "Common/GLUtils.hpp"
 
 using namespace Poly;
