+++ PolyEngine/RenderingDevice/OpenGL/Src/Proxy/GLParticleDeviceProxy.hpp
@@ -2,7 +2,7 @@
 
 #include <Rendering/IRenderingDevice.hpp>
 #include <Utils/EnumUtils.hpp>
-#include "Proxy/GLUtils.hpp"
+#include "Common/GLUtils.hpp"
 
 namespace Poly
 {
