+++ PolyEngine/RenderingDevice/OpenGL/Src/Proxy/GLShaderProgram.cpp
@@ -1,5 +1,5 @@
 #include "Proxy/GLShaderProgram.hpp"
-#include "Proxy/GLUtils.hpp"
+#include "Common/GLUtils.hpp"
 
 SILENCE_MSVC_WARNING(4805, "Warning originates in std::regex");
 #include <regex>
